refactor: remove Container component
