--- React/panda-market/src/components/Layout/Container.jsx
@@ -1,7 +1,0 @@
-import styles from "./Container.module.css";
-
-function Container({ children }) {
-  return <div className={styles.container}>{children}</div>;
-}
-
-export default Container;
