--- React/panda-market/src/components/Layout/Container.module.css
@@ -1,5 +1,0 @@
-.container {
-  max-width: 120rem;
-  margin: 0 auto;
-  width: 100%;
-}
