+++ React/panda-market/src/pages/MarketPage/components/AllItemsSection.module.css
@@ -1,3 +1,7 @@
+.container {
+  padding: 3rem 0;
+}
+
 .top {
   display: flex;
   justify-content: space-between;
@@ -94,6 +98,10 @@ form {
 
 /* Tablet */
 @media (min-width: 768px) and (max-width: 1199px) {
+  .container {
+    padding: 3rem 2.4rem;
+  }
+
   .itemList {
     grid-template-columns: repeat(3, 1fr);
   }
@@ -101,6 +109,10 @@ form {
 
 /* Mobile */
 @media (min-width: 375px) and (max-width: 767px) {
+  .container {
+    padding: 3rem 1.2rem;
+  }
+
   .menu {
     flex-wrap: wrap;
   }
