+++ React/panda-market/src/pages/MarketPage/components/BestItemsSection.module.css
@@ -1,3 +1,7 @@
+.container {
+  padding: 3rem 0;
+}
+
 .title {
   font-weight: 700;
   font-size: 2rem;
@@ -16,6 +20,10 @@
 
 /* Tablet */
 @media (min-width: 768px) and (max-width: 1199px) {
+  .container {
+    padding: 3rem 2.4rem;
+  }
+
   .itemList {
     grid-template-columns: repeat(2, 1fr);
   }
@@ -23,6 +31,10 @@
 
 /* Mobile */
 @media (min-width: 375px) and (max-width: 767px) {
+  .container {
+    padding: 3rem 1.2rem;
+  }
+
   .itemList {
     grid-template-columns: repeat(1, 1fr);
     gap: 0.5rem;
